Refactor or do clauses

+++ lib/perl/Genome/Model/ClinSeq/Command/GenerateClonalityPlots.t
@@ -56,12 +56,11 @@ for my $pdf_name (
 # The differences test excludes files which always differ (embed dates, or are the subject of a masking as above).
 my $temp_dir = "/tmp/last-generate-clonality-plots-result/";
 my @diff     = `diff -x '*.pdf' -x '*.R' -r $expected_out $actual_out`;
-is(scalar(@diff), 0, "only expected differences")
-    or do {
+my $is = is(scalar(@diff), 0, "only expected differences");
+unless ($is) {
     for (@diff) {diag($_)}
     warn
         "*** if the above differences are not in error, rebuild the test data by running this test with REBUILD on the command-line after updating the expected dir's date ***";
     Genome::Sys->shellcmd(cmd => "rm -fr $temp_dir");
     Genome::Sys->shellcmd(cmd => "mv $actual_out $temp_dir");
-    };
-
+}
